+++ assets/css/styles.css
@@ -1,0 +1,292 @@
+body {
+    background-color: #f8f9fa;
+}
+
+.template-preview {
+    border: 1px solid #e9ecef;
+    padding: 15px;
+    margin-top: 15px;
+    border-radius: 8px;
+    display: none;
+    box-shadow: 0 2px 4px rgba(0,0,0,0.05);
+    background: white;
+}
+
+.template-preview img {
+    max-width: 100%;
+    height: auto;
+    border-radius: 6px;
+}
+
+.template-preview.active {
+    display: block;
+    animation: fadeIn 0.3s ease-in;
+}
+
+.form-control:focus {
+    border-color: #4f46e5;
+    box-shadow: 0 0 0 0.2rem rgba(79, 70, 229, 0.25);
+}
+
+.btn-primary {
+    background-color: #4f46e5;
+    border-color: #4f46e5;
+}
+
+.btn-primary:hover {
+    background-color: #4338ca;
+    border-color: #4338ca;
+}
+
+.btn-success {
+    background-color: #10b981;
+    border-color: #10b981;
+}
+
+.btn-success:hover {
+    background-color: #059669;
+    border-color: #059669;
+}
+
+.btn-danger {
+    background-color: #ef4444;
+    border-color: #ef4444;
+}
+
+.btn-danger:hover {
+    background-color: #dc2626;
+    border-color: #dc2626;
+}
+
+@keyframes fadeIn {
+    from { opacity: 0; }
+    to { opacity: 1; }
+}
+
+/* Live Preview Styles */
+#livePreview {
+    position: fixed;
+    top: 20px;
+    right: 20px;
+    width: 340px;
+    background: white;
+    border-radius: 16px;
+    box-shadow: 0 4px 12px rgba(0,0,0,0.15);
+    padding: 0;
+    z-index: 1000;
+    overflow: hidden;
+}
+
+.preview-header {
+    background-color: #262262;
+    padding: 16px 20px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.preview-header img {
+    height: 24px;
+    object-fit: contain;
+}
+
+.preview-header h3 {
+    margin: 0;
+    color: white;
+    font-size: 16px;
+    font-weight: 400;
+}
+
+.preview-body {
+    padding: 16px 20px;
+}
+
+.preview-body-title {
+    font-size: 20px;
+    font-weight: 600;
+    color: #262262;
+    margin-bottom: 4px;
+}
+
+.preview-body-date {
+    font-size: 14px;
+    color: #666;
+    margin-bottom: 16px;
+}
+
+.preview-item {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 12px;
+}
+
+.preview-item-label {
+    color: #666;
+    font-size: 14px;
+}
+
+.preview-item-value {
+    color: #262262;
+    font-weight: 500;
+    font-size: 14px;
+}
+
+.preview-footer {
+    padding: 16px 20px;
+}
+
+.preview-button {
+    display: block;
+    width: 100%;
+    padding: 12px;
+    background: #262262;
+    color: white;
+    text-align: center;
+    border-radius: 8px;
+    text-decoration: none;
+    font-size: 14px;
+    font-weight: 500;
+}
+
+/* Preview Size Controls */
+#livePreview.default { width: 300px; }
+#livePreview.nano { width: 200px; }
+#livePreview.micro { width: 250px; }
+#livePreview.kilo { width: 300px; }
+#livePreview.mega { width: 350px; }
+#livePreview.giga { width: 400px; }
+
+/* Form Styles */
+.form-section {
+    background: #ffffff;
+    border-radius: 1rem;
+    padding: 2rem;
+    margin-bottom: 1.5rem;
+    box-shadow: 0 1px 3px rgba(0,0,0,0.1);
+    transition: all 0.3s ease;
+}
+
+.form-section:hover {
+    box-shadow: 0 4px 6px rgba(0,0,0,0.1);
+    transform: translateY(-2px);
+}
+
+.section-header {
+    display: flex;
+    align-items: center;
+    margin-bottom: 1.5rem;
+    padding-bottom: 0.75rem;
+    border-bottom: 2px solid #e5e7eb;
+}
+
+.section-icon {
+    width: 2rem;
+    height: 2rem;
+    margin-right: 0.75rem;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background: #4f46e5;
+    color: white;
+    border-radius: 0.5rem;
+}
+
+/* Input Fields */
+.form-control, .form-select {
+    border: 1px solid #e5e7eb;
+    border-radius: 0.5rem;
+    padding: 0.75rem;
+    font-size: 0.95rem;
+    transition: all 0.2s;
+    background-color: #f9fafb;
+}
+
+.form-control:focus, .form-select:focus {
+    background-color: #ffffff;
+    border-color: #6366f1;
+    box-shadow: 0 0 0 3px rgba(99, 102, 241, 0.2);
+}
+
+/* Labels */
+.form-label {
+    font-weight: 500;
+    color: #374151;
+    margin-bottom: 0.5rem;
+    font-size: 0.95rem;
+}
+
+/* Buttons */
+.btn {
+    font-weight: 500;
+    padding: 0.75rem 1.5rem;
+    border-radius: 0.5rem;
+    transition: all 0.2s;
+}
+
+.btn-primary {
+    background: linear-gradient(135deg, #4f46e5, #6366f1);
+    border: none;
+}
+
+.btn-primary:hover {
+    background: linear-gradient(135deg, #4338ca, #4f46e5);
+    transform: translateY(-1px);
+}
+
+/* Loading Animation */
+.loading-indicator {
+    width: 100%;
+    height: 3px;
+    background: linear-gradient(90deg, #4f46e5 0%, #6366f1 100%);
+    position: fixed;
+    top: 0;
+    left: 0;
+    z-index: 1000;
+    animation: loading 2s infinite;
+}
+
+@keyframes loading {
+    0% { transform: translateX(-100%); }
+    100% { transform: translateX(100%); }
+}
+
+/* Preview Controls */
+#livePreview {
+    transform: translateX(0);
+}
+
+#livePreview.hidden-preview {
+    transform: translateX(calc(100% + 20px));
+}
+
+#togglePreview {
+    background: transparent;
+    border: none;
+    cursor: pointer;
+    opacity: 0.8;
+}
+
+#togglePreview:hover {
+    opacity: 1;
+}
+
+#showPreviewBtn {
+    transition: all 0.3s ease;
+}
+
+#showPreviewBtn:hover {
+    transform: scale(1.05);
+}
+
+/* Responsive Styles */
+@media (max-width: 768px) {
+    .form-section {
+        padding: 1.5rem;
+    }
+
+    .btn {
+        width: 100%;
+        margin-bottom: 0.5rem;
+    }
+}
