+++ config/config.php
@@ -1,0 +1,20 @@
+<?php
+session_start();
+
+// กำหนดค่าเริ่มต้นต่างๆ
+define('DEFAULT_HEADER_TITLE', 'สำหรับผู้บริหารระดับสูง');
+define('DEFAULT_LOGO_URL', 'https://img5.pic.in.th/file/secure-sv1/JIB_white_logo.png');
+define('DEFAULT_BG_COLOR', '#262262');
+define('DEFAULT_TEXT_COLOR', '#FFFFFF');
+define('DEFAULT_TITLE_SIZE', 'md');
+define('DEFAULT_BODY_TITLE', 'รายงานยอดประจำวัน');
+define('DEFAULT_BODY_TITLE_SIZE', 'xxl');
+define('DEFAULT_BODY_TITLE_COLOR', '#262262');
+define('DEFAULT_BODY_TEXT_COLOR', '#262262');
+define('DEFAULT_BODY_LABEL_SIZE', 'sm');
+define('DEFAULT_BODY_VALUE_SIZE', 'lg');
+define('DEFAULT_FORM_SIZE', 'giga');
+define('DEFAULT_FOOTER_BUTTON_LABEL', 'ดูรายละเอียดเพิ่ม');
+define('DEFAULT_FOOTER_BUTTON_URL', 'https://linecorp.com');
+define('DEFAULT_FOOTER_BUTTON_COLOR', '#262262');
+define('DEFAULT_FOOTER_BUTTON_TEXT_SIZE', 'md');
