+++ includes/footer.php
@@ -1,0 +1,5 @@
+<script src="assets/js/utils.js"></script>
+    <script src="assets/js/preview.js"></script>
+    <script src="assets/js/form.js"></script>
+</body>
+</html>
